pyproject.toml
[project]
name = "video-editing-agent"
version = "0.1.0"
description = "Phase 1 POC for prompt-guided video to ad editing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 120
